connecting splitter to the compilationsuit of arcan
