+++ prover/protocol/compiler/splitter/common.go
