+++ prover/protocol/compiler/splitter/splitter/splitter.go
@@ -2,7 +2,7 @@ package splitter
 
 import (
 	"github.com/consensys/linea-monorepo/prover/protocol/column"
-	alliance "github.com/consensys/linea-monorepo/prover/protocol/compiler/splitter/splitter"
+	alliance "github.com/consensys/linea-monorepo/prover/protocol/compiler/splitter"
 	"github.com/consensys/linea-monorepo/prover/protocol/ifaces"
 	"github.com/consensys/linea-monorepo/prover/protocol/wizard"
 	"github.com/consensys/linea-monorepo/prover/utils"
